requirements.txt
aiogram
python-dotenv
requests
torch
numpy
ultralytics
opencv-python
rembg
onnxruntime
fpdf